backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sihyun Kim <k4sud0n@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
black = "^24.10.0"
django-ninja = "^1.3.0"
django-cors-headers = "^4.6.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
